requirements.txt
APScheduler==3.9.1 # (required)
CherryPy==18.8.0 # (optional)
fritzconnection==1.10.3 # (optional, used by fritzuplink service)
python-gammu==3.2.4 # (optional, used by the sms notification)
paramiko==2.11.0 # (optional, used by ssh service)
pysnmp==4.4.12 # (optional, used by snmp services)
requests==2.28.1 # (optional, used by speedtest service)
xmpp2==0.4 # (optional, used by the xmpp notification)